cogs/requirements.txt
animec==0.3.3
async-timeout==3.0.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
bs4==0.0.1
discord==1.7.3
discord-py-slash-command==2.4.0
discord.py==1.7.3
DiscordUtils==1.3.4
ffmpeg==1.4
ffmpeg-python==0.2.0
lxml==4.5.0
NHentai-API==0.0.16
Pillow==8.3.1
PyNaCl==1.3.0
Quart==0.15.1
Quart-Discord==2.1.4
regex==2021.7.6
wikipedia==1.4.0
wolframalpha==5.0.0
youtube-dl==2021.6.6
chess.com==1.7.5
